citrixbleed-requirements.txt
asn1crypto==1.5.1
asttokens==2.4.1
cffi==1.16.0
chardet==5.2.0
decorator==5.1.1
defusedxml==0.7.1
dissect==3.13
dissect.btrfs==1.2
dissect.cim==3.8
dissect.clfs==1.7
dissect.cstruct==3.13
dissect.esedb==3.12
dissect.etl==3.8
dissect.eventlog==3.7
dissect.evidence==3.8
dissect.executable==1.5
dissect.extfs==3.9
dissect.fat==3.8
dissect.ffs==3.7
dissect.hypervisor==3.12
dissect.jffs==1.1
dissect.ntfs==3.9
dissect.ole==3.7
dissect.regf==3.9
dissect.shellitem==3.7
dissect.sql==3.8
dissect.squashfs==1.4
dissect.target==3.16
dissect.thumbcache==1.7
dissect.util==3.15
dissect.vmfs==3.7
dissect.volume==3.9
dissect.xfs==3.8
exceptiongroup==1.2.1
executing==2.0.1
flow.record==3.14
fusepy==3.0.1
greenlet==0.4.13
hpy==0.9.0
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
msgpack==1.0.8
numpy==1.26.4
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.20.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
readline==6.2.4.1
six==1.16.0
stack-data==0.6.3
structlog==24.1.0
tabulate==0.9.0
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
zstandard==0.22.0
